# docs/API_SPEC.yaml
openapi: 3.1.0
info:
  title: ShiftOps API
  version: 0.1.0
  description: |
    Operational shift control API for HoReCa, accessed by the Telegram Web
    App and the bot webhook handler. The runtime spec is auto-generated by
    FastAPI at `/docs` — that runtime OpenAPI (and the code behind it) is the
    single source of truth. This file is a human-curated overview and may lag.
servers:
  - url: https://shiftops.example.com/api/v1
    description: Production
  - url: http://localhost:8000/api/v1
    description: Local dev

paths:
  /auth/exchange:
    post:
      summary: Exchange Telegram initData for a JWT pair.
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TelegramAuthRequest"
      responses:
        "200":
          description: JWT pair + user payload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TelegramAuthResponse"
        "401":
          description: Invalid / expired initData.

  /shifts/me:
    get:
      summary: Current operator shift + tasks.
      tags: [shifts]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Current shift.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentShiftResponse"
        "404":
          description: No active or scheduled shift.

  /shifts/{shift_id}/start:
    post:
      summary: Start a scheduled shift.
      tags: [shifts]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: shift_id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Shift started.
        "409":
          description: Shift not in `scheduled` status.

  /shifts/tasks/{task_id}/complete:
    post:
      summary: Complete a task (with optional photo).
      tags: [shifts]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: task_id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                comment: { type: string }
                photo: { type: string, format: binary }
      responses:
        "200":
          description: Task marked done.
        "400":
          description: Validation failure (missing required photo, etc.).

  /shifts/tasks/{task_id}/waiver:
    post:
      summary: Request waiver for a critical task.
      tags: [shifts]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: task_id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason: { type: string, enum: [broken, no_staff, no_stock, other] }
                description: { type: string }
      responses:
        "202":
          description: Waiver pending admin decision.

  /shifts/{shift_id}/close:
    post:
      summary: Close the shift (hard/soft block applies).
      tags: [shifts]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: shift_id
          required: true
          schema: { type: string, format: uuid }
        - in: query
          name: confirm_violations
          schema: { type: boolean, default: false }
      responses:
        "200":
          description: Shift closed.
        "409":
          description: Critical tasks pending (hard block) or violations not confirmed.

  /media/{attachment_id}:
    get:
      summary: Proxy redirect to a temporary attachment URL.
      tags: [media]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: attachment_id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "302":
          description: Redirect to the storage provider URL.
        "404":
          description: Not found / unauthorised (existence not leaked).

  /telegram/webhook:
    post:
      summary: Telegram bot webhook.
      tags: [telegram]
      security: []
      parameters:
        - in: header
          name: X-Telegram-Bot-Api-Secret-Token
          required: true
          schema: { type: string }
      responses:
        "200": { description: ok }
        "401": { description: invalid secret }

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    TelegramAuthRequest:
      type: object
      required: [init_data]
      properties:
        init_data: { type: string }

    TelegramAuthResponse:
      type: object
      required: [access_token, refresh_token, token_type, expires_in, user]
      properties:
        access_token: { type: string }
        refresh_token: { type: string }
        token_type: { type: string, example: Bearer }
        expires_in: { type: integer }
        user:
          type: object
          properties:
            id: { type: string, format: uuid }
            tg_user_id: { type: integer }
            full_name: { type: string }
            role: { type: string, enum: [owner, admin, operator] }
            organization_id: { type: string, format: uuid }

    CurrentShiftResponse:
      type: object
      properties:
        shift:
          type: object
          properties:
            id: { type: string, format: uuid }
            template_name: { type: string }
            status:
              type: string
              enum: [scheduled, active, closed_clean, closed_with_violations, aborted]
            score: { type: number, nullable: true }
            progress_done: { type: integer }
            progress_total: { type: integer }
        tasks:
          type: array
          items:
            type: object
            properties:
              id: { type: string, format: uuid }
              title: { type: string }
              description: { type: string, nullable: true }
              criticality: { type: string, enum: [critical, required, optional] }
              requires_photo: { type: boolean }
              status:
                type: string
                enum: [pending, done, skipped, waived, waiver_pending, waiver_rejected]
              has_attachment: { type: boolean }
              completed_at: { type: string, format: date-time, nullable: true }